cleaned a bunch of stuff up, refactored
