--- app/src/main/java/temporal/inventory/receiptsusecase/EmbassyAcknowledgeDataActivity.java
@@ -1,10 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-import io.temporal.activity.ActivityInterface;
-import io.temporal.activity.ActivityMethod;
-
-@ActivityInterface
-public interface EmbassyAcknowledgeDataActivity {
-    @ActivityMethod
-     String ackEvents(String eventData);
-}
