--- app/src/main/java/temporal/inventory/receiptsusecase/EmbassyTransformValidateDataActivity.java
@@ -1,14 +1,0 @@
-package temporal.inventory.receiptsusecase;
-import io.temporal.activity.ActivityInterface;
-import io.temporal.activity.ActivityMethod;
-
-    
-@ActivityInterface
-    
-public interface EmbassyTransformValidateDataActivity {
-    @ActivityMethod
-    String processRecord(String eventType);
-    @ActivityMethod
-    String validateRecord(String eventType);
-}
-
