--- app/src/main/java/temporal/inventory/receiptsusecase/GEOTransformModelActivity.java
@@ -1,11 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-import io.temporal.activity.ActivityInterface;
-import io.temporal.activity.ActivityMethod;
-
-@ActivityInterface
-public interface GEOTransformModelActivity {
-    @ActivityMethod
-    String tranformtoeventmodel();
-    
-}
