--- app/src/main/java/temporal/inventory/receiptsusecase/GEOTransformModelActivityImpl.java
@@ -1,14 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-
-import org.springframework.stereotype.Component;
-
-@Component("geo-transform-model-activity")
-public class GEOTransformModelActivityImpl implements GEOTransformModelActivity {
-    @Override
-    public String tranformtoeventmodel(){
-     
-        return "Inventory DataModel transfromed to Event Model:";
-
-    }
-}
