--- app/src/main/java/temporal/inventory/receiptsusecase/PublishActivity.java
@@ -1,11 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-import io.temporal.activity.ActivityInterface;
-import io.temporal.activity.ActivityMethod;
-
-@ActivityInterface
-public interface PublishActivity {
-    @ActivityMethod
-    String publishEvents();
-    
-}
