--- app/src/main/java/temporal/inventory/receiptsusecase/PublishActivityImpl.java
@@ -1,12 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-
-public class PublishActivityImpl implements PublishActivity{
-    
-   @Override
-    public String publishEvents(){
-    
-     return "TRANSFER RECEIPTS Event Data Published:";
-
-    }
-}
