--- app/src/main/java/temporal/inventory/receiptsusecase/SaveStatusActivity.java
@@ -1,13 +1,0 @@
-
-package temporal.inventory.receiptsusecase;
-import io.temporal.activity.ActivityInterface;
-import io.temporal.activity.ActivityMethod;
-
-@ActivityInterface
-public interface SaveStatusActivity {
-    @ActivityMethod
-    String savestatus(String status);
-
-}
-
-
